pyproject.toml
[tool.poetry]
name = "fastapi-mongodb"
version = "0.1.0"
description = ""
authors = ["shahzad.ahmad <shahzad@renesistech.com>"]
readme = "README.md"
package-mode = false

packages = [
    { include = "cmd" }
]



[tool.poetry.dependencies]
python = "^3.11"
annotated-types = "==0.7.0"
anyio = "==4.4.0"
certifi = "==2024.6.2"
click = "==8.1.7"
dnspython = "==2.6.1"
email-validator = "==2.2.0"
fastapi = "==0.111.0"
fastapi-cli = "==0.0.4"
h11 = "==0.14.0"
httpcore = "==1.0.5"
httptools = "==0.6.1"
httpx = "==0.27.0"
idna = "==3.7"
Jinja2 = "==3.1.4"
markdown-it-py = "==3.0.0"
MarkupSafe = "==2.1.5"
mdurl = "==0.1.2"
motor = "==3.5.0"
orjson = "==3.10.5"
pydantic = "==2.8.0"
pydantic-core = "==2.20.0"
Pygments = "==2.18.0"
pymongo = "==4.8.0"
python-dotenv = "==1.0.1"
python-multipart = "==0.0.9"
PyYAML = "==6.0.1"
rich = "==13.7.1"
shellingham = "==1.5.4"
sniffio = "==1.3.1"
starlette = "==0.37.2"
typer = "==0.12.3"
typing-extensions = "==4.12.2"
ujson = "==5.10.0"
uvicorn = "==0.30.1"
uvloop = "==0.19.0"
watchfiles = "==0.22.0"
websockets = "==12.0"
passlib = "==1.7.4"
bcrypt = "==4.0.1"
fastapi-sso = "==0.15.0"
cryptography = "==42.0.8"
aiosmtplib = "==3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.black]
line-length = 79
target-version = ['py37']  # Adjust to match your Python version requirements
include = '\.pyi?$'  # Include .py and .pyi files

[tool.poetry.scripts]
runserver = "python3 -B run.py"
format = "black app/"
